chore: move the component's hook to the root hook folder
