+++ src/components/Header/Header.jsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import useHeader from '../../hooks/for-components/useHeader';
+import useHeader from '../../hooks/useHeader';
 import HeaderNavigation from '../HeaderNavigation/HeaderNavigation';
 import HeaderMenuButton from '../HeaderMenuButton/HeaderMenuButton';
 import headerLogoUrl from '../../images/header-logo.svg';
