+++ src/components/HistoryFactList/HistoryFactList.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import CustomList from '../CustomList/CustomList';
 import HistoryFactItems from '../HistoryFactItems/HistoryFactItems';
-import useHistory from '../../hooks/for-components/useHistory';
+import useHistory from '../../hooks/useHistory';
 import TEXT_CONTENT from '../../utils/scripts/text-content';
 
 function HistoryFactList({ selectedId, styles }) {
